fix: more possible change deadlocks
