+++ launchpad/handlers/http.go
@@ -1,16 +1,12 @@
 package handlers
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
-	"openfeature.com/flagd-testbed/launchpad/pkg"
-	"os"
 	"strconv"
-	"sync"
-	"time"
+
+	flagd "openfeature.com/flagd-testbed/launchpad/pkg"
 )
 
 // Response struct to standardize API responses
@@ -34,7 +30,7 @@ func StartFlagdHandler(w http.ResponseWriter, r *http.Request) {
 func RestartHandler(w http.ResponseWriter, r *http.Request) {
 	secondsStr := r.URL.Query().Get("seconds")
 	if secondsStr == "" {
-		secondsStr = "5"
+		secondsStr = fmt.Sprintf("%d", flagd.DefaultRestartTimeout)
 	}
 
 	seconds, err := strconv.Atoi(secondsStr)
@@ -56,88 +52,14 @@ func StopFlagdHandler(w http.ResponseWriter, r *http.Request) {
 	respondWithJSON(w, http.StatusOK, "success", "flagd stopped successfully")
 }
 
-type FlagConfig struct {
-	Flags map[string]struct {
-		State          string            `json:"state"`
-		Variants       map[string]string `json:"variants"`
-		DefaultVariant string            `json:"defaultVariant"`
-	} `json:"flags"`
-}
-
-var mu sync.Mutex // Mutex to ensure thread-safe file operations
-
 // ChangeHandler triggers JSON file merging and notifies `flagd`
 func ChangeHandler(w http.ResponseWriter, r *http.Request) {
-	mu.Lock()
-	defer mu.Unlock()
-
-	// Path to the configuration file
-	configFile := "rawflags/changing-flag.json"
-
-	// Read the existing file
-	data, err := os.ReadFile(configFile)
+	variant, err := flagd.ToggleChangingFlag()
 	if err != nil {
-		http.Error(w, fmt.Sprintf("Failed to read file: %v", err), http.StatusInternalServerError)
+		http.Error(w, fmt.Sprintf("Failed to toggle changing flag: %v", err), http.StatusInternalServerError)
 		return
 	}
-
-	// Parse the JSON into the FlagConfig struct
-	var config FlagConfig
-	if err := json.Unmarshal(data, &config); err != nil {
-		http.Error(w, fmt.Sprintf("Failed to parse JSON: %v", err), http.StatusInternalServerError)
-		return
-	}
-
-	// Find the "changing-flag" and toggle the default variant
-	flag, exists := config.Flags["changing-flag"]
-	if !exists {
-		http.Error(w, "Flag 'changing-flag' not found in the configuration", http.StatusNotFound)
-		return
-	}
-
-	// Toggle the defaultVariant between "foo" and "bar"
-	if flag.DefaultVariant == "foo" {
-		flag.DefaultVariant = "bar"
-	} else {
-		flag.DefaultVariant = "foo"
-	}
-
-	// Save the updated flag back to the configuration
-	config.Flags["changing-flag"] = flag
-	// Serialize the updated configuration back to JSON
-	updatedData, err := json.MarshalIndent(config, "", "  ")
-	if err != nil {
-		http.Error(w, fmt.Sprintf("Failed to serialize updated JSON: %v", err), http.StatusInternalServerError)
-		return
-	}
-
-	// the file watcher should be triggered instantly. If not, we add a timeout to prevent a hanging test
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	// wait for the filewatcher to register an update and write the new json file
-	flagUpdateWait := sync.WaitGroup{}
-	flagUpdateWait.Add(1)
-	flagd.RegisterWaitForNextChangingFlagUpdate(&flagUpdateWait)
-	go func() {
-		flagUpdateWait.Wait()
-		cancel()
-	}()
-
-	// Write the updated JSON back to the file
-	if err := os.WriteFile(configFile, updatedData, 0644); err != nil {
-		http.Error(w, fmt.Sprintf("Failed to write updated file: %v", err), http.StatusInternalServerError)
-		return
-	}
-
-	select {
-	case <-ctx.Done():
-		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
-			http.Error(w, fmt.Sprintf("Flags were not updated in time: %v", ctx.Err()), http.StatusInternalServerError)
-		} else {
-			respondWithJSON(w, http.StatusOK, "success", fmt.Sprintf("Default variant successfully changed to '%s'\n", flag.DefaultVariant))
-		}
-	}
+	respondWithJSON(w, http.StatusOK, "success", fmt.Sprintf("Default variant successfully changed to '%s'\n", variant))
 }
 
 // Utility function to send JSON responses
